requirements.txt
requests==2.32.3
pandas==2.2.3
pyarrow==19.0.1
google-cloud-storage==2.19.0
google-cloud-bigquery==3.29.0
python-dotenv==1.0.1
tenacity==9.0.0
functions-framework==3.8.2 #pour que cloud function reconnaisse le point d'entrée HTTP